+++ backend/src/test/java/com/woowacourse/thankoo/coupon/domain/CouponHistoryQueryRepositoryTest.java
@@ -2,9 +2,9 @@ package com.woowacourse.thankoo.coupon.domain;
 
 import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HOHO;
 import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI;
-import static com.woowacourse.thankoo.common.fixtures.TestFixture.MESSAGE;
-import static com.woowacourse.thankoo.common.fixtures.TestFixture.TITLE;
-import static com.woowacourse.thankoo.common.fixtures.TestFixture.TYPE;
+import static com.woowacourse.thankoo.common.fixtures.CouponFixture.MESSAGE;
+import static com.woowacourse.thankoo.common.fixtures.CouponFixture.TITLE;
+import static com.woowacourse.thankoo.common.fixtures.CouponFixture.TYPE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.woowacourse.thankoo.member.domain.Member;
